requirements.txt
aniso8601==9.0.1
antlr4-python3-runtime==4.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
bleach==4.1.0
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
Cython==3.0.0a9
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.2
entrypoints==0.3
executing==0.8.2
ffmpeg==1.4
fire==0.4.0
Flask==2.0.2
Flask-RESTful==0.3.9
fonttools==4.29.1
future==0.18.2
fvcore==0.1.5.post20220119
hydra-core==1.1.1
idna==3.3
image==1.5.33
imageio==2.14.1
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==4.8.2
iopath==0.1.9
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
labelme2coco==0.2.1
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
numpy==1.21.4
omegaconf==2.1.1
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
pixellib==0.7.1
platformdirs==2.4.1
portalocker==2.3.2
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.1
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==1.11.2
requests==2.27.1
s3transfer==0.5.0
sahi==0.8.22
scikit-image==0.18.3
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.4.2
Shapely==1.8.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tabulate==0.8.9
termcolor==1.1.0
terminado==0.13.1
terminaltables==3.1.10
testpath==0.5.0
tifffile==2021.11.2
tomli==2.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
xmljson==0.2.1
yacs==0.1.8
